Fix the README.md to remove reference to .gitignore
Will not add TMDbKey.java to the .gitignore, since the TMDbKey.java
file with the “PLACE API KEY HERE” should be populated at other users’
projects.  It is important to remember the —assume-unchanged once the
real key replaces the “PLACE API KEY HERE”.  By the way, at this time
the real key has been set in the TMDbKey.java file.  But that file will
be —assume-unchanged